Update anti_propkill.lua
Added Debug menu for me to test things easier.  This is only accessible by me and requires me to be superadmin on the server that I run the command so don't worry about me joining and messing everything up.  Currently it just allows me to run a test message and it will never allow me to do anything malicious as that's just messed up.  

Fixed vehicles changing colors when hitting someone so now it just 'freezes' it even though enabling motion of the vehicle does nothing.  This should be fixed in the next update.

Fixed vehicles not sending a message to staff (Completely forgot to add the message for it :/).

Added a way to freeze players with the physgun (this will soon be toggle-able when the next update comes out)  with messages when someone freezes a person and unfreezes them.  It also gives them god to prevent them to be killed while in it.

Plans for next update:
- Fix vehicles allowing people to kill people if they are against a wall (I might make it so the vehicle gets no-collided from players only for 5 seconds then if the vehicle is still inside someone it will wait another and so on till it's not in a player and can re-collide with players.)
- Add a config to allow the enabling and disabling of certain parts of the addon.
- Add a menu to allow changing the config in-game by a specific group defined in the config (default: Superadmin)
This is assuming I don't have to put out a hotfix to a huge problem.